enhancement/allow-intermediate-service-to-continue
-running-when-child-error

+++ octue/cloud/pub_sub/service.py
@@ -8,11 +8,11 @@ import math
 import time
 import uuid
 
+from google import auth
 from google.api_core import retry
 from google.cloud import pubsub_v1
 
 import octue.exceptions
-from octue.cloud.credentials import GCPCredentialsManager
 from octue.cloud.pub_sub import Subscription, Topic
 from octue.cloud.pub_sub.logging import GooglePubSubHandler
 from octue.cloud.pub_sub.message_handler import OrderedMessageHandler
@@ -60,8 +60,7 @@ class Service(CoolNameable):
 
         self.backend = backend
         self.run_function = run_function
-
-        self._credentials = GCPCredentialsManager(backend.credentials_environment_variable).get_credentials()
+        self._credentials = auth.default()[0]
         self.publisher = pubsub_v1.PublisherClient(credentials=self._credentials, batch_settings=BATCH_SETTINGS)
         self._current_question = None
         super().__init__(*args, **kwargs)
@@ -77,11 +76,9 @@ class Service(CoolNameable):
         :param bool delete_topic_and_subscription_on_exit: if `True`, delete the service's topic and subscription on exit
         :return None:
         """
-        logger.info("%r started with ID %r.", self, self.id)
+        logger.info("Starting service with ID %r.", self.id)
 
         topic = Topic(name=self.id, namespace=OCTUE_NAMESPACE, service=self)
-        topic.create(allow_existing=True)
-
         subscriber = pubsub_v1.SubscriberClient(credentials=self._credentials)
 
         subscription = Subscription(
@@ -92,15 +89,20 @@ class Service(CoolNameable):
             subscriber=subscriber,
             expiration_time=None,
         )
-        subscription.create(allow_existing=True)
 
-        start_time = time.perf_counter()
-        run_time = 0
+        try:
+            topic.create(allow_existing=True)
+            subscription.create(allow_existing=True)
 
-        if timeout is None:
-            timeout = math.inf
+            future = subscriber.subscribe(subscription=subscription.path, callback=self.answer)
+            logger.debug("%r is waiting for questions.", self)
 
-        with subscriber:
+            start_time = time.perf_counter()
+            run_time = 0
+
+            if timeout is None:
+                timeout = math.inf
+
             while run_time < timeout:
 
                 logger.info("%r is waiting for new questions...", self)
@@ -122,9 +124,21 @@ class Service(CoolNameable):
                     future.cancel()
                     break
 
+        finally:
             if delete_topic_and_subscription_on_exit:
-                topic.delete()
-                subscription.delete()
+                try:
+                    if subscription.exists():
+                        subscription.delete()
+                        logger.info("Subscription deleted.")
+
+                    if topic.exists():
+                        topic.delete()
+                        logger.info("Topic deleted.")
+
+                except Exception:
+                    logger.error("Deletion of topic and/or subscription %r failed.", topic.name)
+
+            subscriber.close()
 
     def answer(self, question, answer_topic=None, timeout=30):
         """Answer a question (i.e. run the Service's app to analyse the given data, and return the output values to the
@@ -200,8 +214,9 @@ class Service(CoolNameable):
         input_manifest=None,
         subscribe_to_logs=True,
         allow_local_files=False,
-        timeout=30,
         question_uuid=None,
+        push_endpoint=None,
+        timeout=30,
     ):
         """Ask a serving Service a question (i.e. send it input values for it to run its app on). The input values must
         be in the format specified by the serving Service's Twine file. A single-use topic and subscription are created
@@ -213,8 +228,9 @@ class Service(CoolNameable):
         :param octue.resources.manifest.Manifest|None input_manifest: the input manifest of the question
         :param bool subscribe_to_logs: if `True`, subscribe to logs from the remote service and handle them with the local log handlers
         :param bool allow_local_files: if `True`, allow the input manifest to contain references to local files - this should only be set to `True` if the serving service will have access to these local files
-        :param float|None timeout: time in seconds to keep retrying sending the question
         :param str|None question_uuid: the UUID to use for the question if a specific one is needed; a UUID is generated if not
+        :param str|None push_endpoint: if answers to the question should be pushed to an endpoint, provide its URL here; if they should be pulled, leave this as `None`
+        :param float|None timeout: time in seconds to keep retrying sending the question
         :return (octue.cloud.pub_sub.subscription.Subscription, str): the response subscription and question UUID
         """
         if not allow_local_files:
@@ -247,6 +263,7 @@ class Service(CoolNameable):
             namespace=OCTUE_NAMESPACE,
             project_name=self.backend.project_name,
             subscriber=pubsub_v1.SubscriberClient(credentials=self._credentials),
+            push_endpoint=push_endpoint,
         )
         response_subscription.create(allow_existing=True)
 
@@ -298,6 +315,11 @@ class Service(CoolNameable):
         :raise TimeoutError: if the timeout is exceeded
         :return dict: dictionary containing the keys "output_values" and "output_manifest"
         """
+        if subscription.is_push_subscription:
+            raise octue.exceptions.PushSubscriptionCannotBePulled(
+                f"Cannot pull from {subscription.path!r} subscription as it is a push subscription."
+            )
+
         subscriber = pubsub_v1.SubscriberClient(credentials=self._credentials)
 
         message_handler = OrderedMessageHandler(
@@ -307,39 +329,38 @@ class Service(CoolNameable):
             service_name=service_name,
         )
 
-        with subscriber:
+        try:
+            # Retry sending the question until the overall timeout is reached.
+            while not message_handler.received_delivery_acknowledgement:
 
-            try:
-                # Retry sending the question until the overall timeout is reached.
-                while not message_handler.received_delivery_acknowledgement:
+                try:
+                    return message_handler.handle_messages(
+                        timeout=timeout,
+                        delivery_acknowledgement_timeout=delivery_acknowledgement_timeout,
+                    )
 
-                    try:
-                        return message_handler.handle_messages(
-                            timeout=timeout,
-                            delivery_acknowledgement_timeout=delivery_acknowledgement_timeout,
-                        )
+                except octue.exceptions.QuestionNotDelivered:
+                    logger.info(
+                        "%r: No acknowledgement of question delivery after %fs - resending in %fs.",
+                        self,
+                        delivery_acknowledgement_timeout,
+                        retry_interval,
+                    )
 
-                    except octue.exceptions.QuestionNotDelivered:
-                        logger.info(
-                            "%r: No acknowledgement of question delivery after %fs - resending in %fs.",
-                            self,
-                            delivery_acknowledgement_timeout,
-                            retry_interval,
-                        )
+                    time.sleep(retry_interval)
+                    self.ask(**self._current_question)
 
-                        time.sleep(retry_interval)
-                        self.ask(**self._current_question)
+                except Exception as e:
 
-                    except Exception as e:
+                    if just_log_errors:
+                        logger.exception(e)
+                        raise octue.exceptions.AbandonAnalysis()
 
-                        if just_log_errors:
-                            logger.exception(e)
-                            raise octue.exceptions.AbandonAnalysis()
+                    raise e
 
-                        raise e
-
-            finally:
-                subscription.delete()
+        finally:
+            subscription.delete()
+            subscriber.close()
 
     def _send_delivery_acknowledgment(self, topic, timeout=30):
         """Send an acknowledgement of question delivery to the asker.
